requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
pandas
plotly
openai
speech_recognition
sounddevice
bcrypt
numpy